move checkin to api
