+++ apps/sspanel/urls.py
@@ -21,7 +21,6 @@ urlpatterns = [
     # 用户信息
     path('users/userinfo/', views.userinfo, name='userinfo'),
     path('users/userinfoedit/', views.userinfo_edit, name='userinfo_edit'),
-    path('checkin/', views.checkin, name='checkin'),
     #  二维码
     path('qrcode/ssr/<node_id>/',
          views.get_ssr_qrcode, name='ssrqrcode'),
